Move AttachmentHolderCodec to dedicated class, fix tests not compiling
